Minor mods. After a stop, the contact is free

+++ cyclo/src/include/program_manager.hpp
@@ -35,14 +35,16 @@ SOFTWARE.
  * Created: 17/07/2021 18:22:20
  *  Author: micro
  */
+#include "contact.hpp"
+#include "parser.hpp"
+#include "program.hpp"
+#include "trace.h"
+
 #include <etl/bitset.h>
 #include <etl/string_view.h>
 
-#include "program.hpp"
 #include "conf_cyclo.hpp"
-#include "contact.hpp"
-#include "parser.hpp"
-#include "trace.h"
+
 
 
 /**
@@ -54,7 +56,7 @@ class ProgramManager
 {
 public:
    /** A bitset which determine which slot contains a valid program */
-   using Pgms                                 = etl::bitset<10>;
+   using Pgms = etl::bitset<10>;
 
    /** The state of the active program  */
    enum program_state_t : uint8_t { stopped, paused, running, usb };
@@ -63,7 +65,7 @@ public:
    enum mode_t : uint8_t { normal, autostart };
 
    /** The storage for the program must fit 2 pages of eeprom - header, spare and crc (2 bytes) */
-   static constexpr size_t STORAGE_MAX_LENGTH = (EEPROM_PAGE_SIZE * 2) - 4;
+   static constexpr size_t STORAGE_MAX_LENGTH = ( EEPROM_PAGE_SIZE * 2 ) - 4;
 
 private:
    ///< Current selected program. 0 is auto. -1 is none.
@@ -76,7 +78,7 @@ private:
    program_state_t state;
 
    /** Current value of the counter. uint16_max for undefined */
-   uint16_t counter;
+   int32_t counter;
 
    /** Bit field containing program slots taken */
    Pgms occupancy_map;
@@ -101,17 +103,17 @@ public:
    //////////////////////////////////////////////////////////////////////////
 public:
    inline int8_t get_selected() { return selected; }
-   inline void    set_selected( int8_t pgm ) { selected = pgm; }
+   inline void   set_selected( int8_t pgm ) { selected = pgm; }
 
    inline program_state_t get_state() { return state; }
    inline void            set_state( program_state_t newState ) { state = newState; }
 
    ///< Grab the execution counter
-   inline uint16_t get_counter() { return counter; }
-   inline void     set_counter( uint16_t newValue ) { counter = newValue; }
+   inline int32_t get_counter() { return counter; }
+   inline void     set_counter( int32_t newValue ) { counter = newValue; }
 
    // @return true if a program starts automatically
-   inline bool starts_automatically() const { return (auto_start >= 0); }
+   inline bool starts_automatically() const { return ( auto_start >= 0 ); }
 
    // @return true if a program starts automatically
    inline int8_t get_autostart_index() const { return auto_start; }
@@ -147,7 +149,7 @@ public:
    void scan();
 
    /** Set a program has auto start */
-   void set_autostart(uint8_t index );
+   void set_autostart( uint8_t index );
 
    //
    // Drive the sequencer
@@ -167,10 +169,7 @@ protected:
       return static_cast<T *>( address );
    }
 
-   eeprom_addr_t pgm_at( const size_t index )
-   {
-      return EEPROM_PAGE_SIZE * index * 2;
-   }
+   eeprom_addr_t pgm_at( const size_t index ) { return EEPROM_PAGE_SIZE * index * 2; }
 };
